refactor: rename LLM module from "datastep" to "llm" & child dirs
