+++ src/llm/chain/datastep_check_data_chain.py
@@ -4,7 +4,7 @@ from langchain.chains import LLMChain
 from langchain.prompts.prompt import PromptTemplate
 from langchain_openai import AzureChatOpenAI
 
-from datastep.components.sql_database import DatastepSqlDatabase
+from llm.component.sql_database import DatastepSqlDatabase
 from infra.env import AZURE_DEPLOYMENT_NAME_DB_ASSISTANT
 from util.logger import async_log
