+++ src/model/mapping/mapping_cim_work_type_model.py
@@ -1,6 +1,6 @@
 import json
 
-from datastep.chains.datastep_cim_mapping_chain import get_chain_for_cim
+from llm.chain.datastep_cim_mapping_chain import get_chain_for_cim
 from scheme.mapping.mapping_results_scheme import InputModel, MappedCimModel
