+++ src/model/order_classification/order_classification_model.py
@@ -5,7 +5,7 @@ import requests
 from fastapi import HTTPException, status
 from loguru import logger
 
-from datastep.chains.order_multi_classification.order_multi_classification_chain import (
+from llm.chain.order_multi_classification.order_multi_classification_chain import (
     get_order_class,
 )
 from infra.env import (
